Feat/user profile (bio, profile image)

+++ src/app/api/user/route.ts
@@ -1,6 +1,9 @@
-import { ApiError, ValidationError } from '@/errors/errors';
+import { auth } from '@/auth';
+import { ApiError, UnauthorizedError, ValidationError } from '@/errors/errors';
 import * as UserQueries from '@/features/user/data/queries/userQueries';
+import * as ProfileUsecase from '@/features/user/data/usecases/profileUsecase';
 import { getUserId } from '@/lib/user';
+import { cookies } from 'next/headers';
 import { NextRequest, NextResponse } from 'next/server';
 import 'server-only';
 
@@ -25,6 +28,13 @@ export async function GET() {
 
 export async function PATCH(request: NextRequest) {
   try {
+    const userId = (await cookies()).get('userId')?.value;
+    const authUserId = (await auth())?.user?.id;
+
+    if (!userId || !authUserId) {
+      throw new UnauthorizedError('인증되지 않은 요청입니다');
+    }
+
     const { nickname } = await request.json();
 
     if (!nickname) {
@@ -35,7 +45,7 @@ export async function PATCH(request: NextRequest) {
       throw new ValidationError('닉네임은 1-50자여야 합니다');
     }
 
-    await UserQueries.updateUser(nickname);
+    await UserQueries.updateUser({ userId, authUserId, nickname });
 
     return NextResponse.json({ data: null });
   } catch (error) {
@@ -54,7 +64,15 @@ export async function PATCH(request: NextRequest) {
 
 export async function DELETE() {
   try {
+    const session = await auth();
+    const userId = session?.user?.user_id;
+
+    if (!userId) {
+      throw new UnauthorizedError('인증되지 않은 요청입니다');
+    }
+
     await Promise.all([
+      ProfileUsecase.deleteProfileMedia(userId),
       UserQueries.deleteUser(),
       UserQueries.hardDeleteAuthUser(),
     ]);
